# src/dscompanion/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dscompanion API",
    "description": "REST surface over dscompanion's 13-step interactive pipeline wizard.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/runs/{run_id}/steps/feature_processing/preview": {
      "post": {
        "tags": [
          "feature_processing"
        ],
        "summary": "Preview",
        "description": "Computes every feature column's recommended recipe, without persisting anything.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    FeatureProcessingPreviewResponse: The full candidate list.\n\nRaises:\n    HTTPException: 400, if Steps 1-3 have not been confirmed for this run.",
        "operationId": "preview_api_runs__run_id__steps_feature_processing_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProcessingPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/feature_processing/transform-preview": {
      "post": {
        "tags": [
          "feature_processing"
        ],
        "summary": "Transform Preview",
        "description": "Fits a real transform chain on the client's current accepted recipes and returns\na before/after sample.\n\nArgs:\n    body (FeatureProcessingTransformPreviewRequest): Current accepted recipes.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    FeatureProcessingTransformPreviewResponse: Before/after sample rows.\n\nRaises:\n    HTTPException: 400, if Steps 1-3 have not been confirmed, or the recipes fail\n        to fit.",
        "operationId": "transform_preview_api_runs__run_id__steps_feature_processing_transform_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureProcessingTransformPreviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProcessingTransformPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/feature_processing/confirm": {
      "post": {
        "tags": [
          "feature_processing"
        ],
        "summary": "Confirm",
        "description": "Persists the final recipe/drop decisions for this run.\n\nArgs:\n    body (FeatureProcessingConfirmRequest): Final accepted recipes and dropped\n        columns.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    FeatureProcessingConfirmResponse: Confirmation and accepted/dropped/rejected\n    counts.\n\nRaises:\n    HTTPException: 400, if Steps 1-3 have not been confirmed for this run.",
        "operationId": "confirm_api_runs__run_id__steps_feature_processing_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeatureProcessingConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeatureProcessingConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/runs/{run_id}/steps/evaluate/preview": {
      "post": {
        "tags": [
          "evaluate"
        ],
        "summary": "Preview",
        "description": "Evaluates the final model from Step 9 and reports the full breakdown.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    EvaluatePreviewResponse: Full per-split metric breakdown, the\n    tuned-vs-default baseline, and any suspicious-pattern flags.\n\nRaises:\n    HTTPException: 400, if Step 9 has not been confirmed for this run.",
        "operationId": "preview_api_runs__run_id__steps_evaluate_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluatePreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/evaluate/confirm": {
      "post": {
        "tags": [
          "evaluate"
        ],
        "summary": "Confirm",
        "description": "Records Step 10 completion \u2014 a flag-flip only, no persisted user choice.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    EvaluateConfirmResponse: Confirmation.\n\nRaises:\n    HTTPException: 400, if Step 9 has not been confirmed for this run.",
        "operationId": "confirm_api_runs__run_id__steps_evaluate_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluateConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/calibration/preview": {
      "post": {
        "tags": [
          "calibration"
        ],
        "summary": "Preview",
        "description": "Fits calibration on the final model and reports ECE before/after.\n\nArgs:\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    CalibrationPreviewResponse: Availability, ECE before/after, and the\n    apply/skip recommendation.\n\nRaises:\n    HTTPException: 400, if Step 9 has not been confirmed, or no\n        held-out split is available, for this run.",
        "operationId": "preview_api_runs__run_id__steps_calibration_preview_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationPreviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/steps/calibration/confirm": {
      "post": {
        "tags": [
          "calibration"
        ],
        "summary": "Confirm",
        "description": "Applies or skips calibration and persists ``calibrated_model``.\n\nArgs:\n    body (CalibrationConfirmRequest): Whether to apply calibration.\n    run (RunState): Injected via ``get_pipeline_service``.\n\nReturns:\n    CalibrationConfirmResponse: Confirmation and whether calibration was\n    actually applied.\n\nRaises:\n    HTTPException: 400, if Step 9 has not been confirmed, or no\n        held-out split is available, for this run.",
        "operationId": "confirm_api_runs__run_id__steps_calibration_confirm_post",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationConfirmResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalibrationConfirmRequest": {
        "properties": {
          "apply": {
            "type": "boolean",
            "title": "Apply"
          }
        },
        "type": "object",
        "required": [
          "apply"
        ],
        "title": "CalibrationConfirmRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/calibration/confirm``.\n\nArgs:\n    apply (bool): ``True`` to wrap the model with calibrated\n        probabilities, ``False`` to keep the raw scores unchanged.\n        Ignored (treated as ``False``) for non-classification tasks."
      },
      "CalibrationConfirmResponse": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied"
          }
        },
        "type": "object",
        "required": [
          "confirmed",
          "applied"
        ],
        "title": "CalibrationConfirmResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/calibration/confirm``.\n\nArgs:\n    confirmed (bool): Always ``True`` on success.\n    applied (bool): Whether calibration was actually applied \u2014 ``False``\n        for a non-classification task regardless of the request."
      },
      "CalibrationPreviewResponse": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "split_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Split Used"
          },
          "ece_before": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ece Before"
          },
          "ece_after": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ece After"
          },
          "recommend_apply": {
            "type": "boolean",
            "title": "Recommend Apply",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "CalibrationPreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/calibration/preview``.\n\nArgs:\n    available (bool): ``False`` for non-classification tasks \u2014 the\n        frontend should auto-skip straight to confirming with\n        ``apply=False``, matching Streamlit's silent auto-skip rather\n        than offering a choice that has no effect.\n    split_used (str | None): Held-out split the calibrator was fit and\n        evaluated on (``\"val\"`` if available, else ``\"test\"``, else\n        ``\"train\"``). ``None`` when ``available=False``.\n    ece_before (float | None): Expected Calibration Error before\n        calibration. ``None`` when ``available=False``.\n    ece_after (float | None): Expected Calibration Error after isotonic\n        calibration. ``None`` when ``available=False``.\n    recommend_apply (bool): ``True`` when calibration meaningfully\n        reduced ECE \u2014 drives which of the two choice buttons the\n        frontend highlights as primary. Always ``False`` when\n        ``available=False``."
      },
...
      "EvaluateConfirmResponse": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          }
        },
        "type": "object",
        "required": [
          "confirmed"
        ],
        "title": "EvaluateConfirmResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/evaluate/confirm``.\n\nArgs:\n    confirmed (bool): Always ``True`` on success."
      },
...
      "EvaluateMetricRow": {
        "properties": {
          "split": {
            "type": "string",
            "title": "Split"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "gloss": {
            "type": "string",
            "title": "Gloss"
          },
          "value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "split",
          "metric",
          "label",
          "gloss"
        ],
        "title": "EvaluateMetricRow",
        "description": "One ``(split, metric)`` pair from Step 10's full metric breakdown.\n\nArgs:\n    split (str): One of ``\"train\"``, ``\"val\"``, ``\"test\"``, ``\"oot\"``.\n    metric (str): Metric key (e.g. ``\"roc_auc\"``).\n    label (str): Plain-language display label (e.g. ``\"AUC-ROC\"``).\n    gloss (str): Plain-language explanation of what the metric means.\n    value (float | None): Metric value, ``None`` if unavailable or ``NaN``\n        for this split."
      },
      "EvaluatePreviewResponse": {
        "properties": {
          "algorithm": {
            "type": "string",
            "title": "Algorithm"
          },
          "tuned": {
            "type": "boolean",
            "title": "Tuned"
          },
          "primary_split": {
            "type": "string",
            "title": "Primary Split"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/EvaluateMetricRow"
            },
            "type": "array",
            "title": "Metrics"
          },
          "baseline": {
            "items": {
              "$ref": "#/components/schemas/EvaluateBaseline"
            },
            "type": "array",
            "title": "Baseline"
          },
          "flags": {
            "items": {
              "$ref": "#/components/schemas/EvaluateFlag"
            },
            "type": "array",
            "title": "Flags"
          }
        },
        "type": "object",
        "required": [
          "algorithm",
          "tuned",
          "primary_split"
        ],
        "title": "EvaluatePreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/evaluate/preview``.\n\nArgs:\n    algorithm (str): The confirmed algorithm from Step 8.\n    tuned (bool): Whether Step 9's tuned model was accepted \u2014 controls\n        whether ``baseline`` is populated.\n    primary_split (str): The split used for headline metrics and flags \u2014\n        ``\"val\"`` if available, else ``\"test\"``, else ``\"train\"``.\n    metrics (list[EvaluateMetricRow]): Full per-split metric breakdown,\n        headline metrics (``roc_auc``, ``gini``, ``ks_statistic``) first,\n        then the remaining supporting metrics.\n    baseline (list[EvaluateBaseline]): Step 8's untuned headline-metric\n        values on ``primary_split``, for the \"vs. default\" delta shown\n        when tuned. Empty when ``tuned=False``.\n    flags (list[EvaluateFlag]): Suspicious-pattern warnings. Empty when\n        none were found."
      },
...
      "FeatureProcessingConfirmRequest": {
        "properties": {
          "recipes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ColumnRecipe"
            },
            "type": "object",
            "title": "Recipes"
          },
          "dropped_columns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Dropped Columns"
          }
        },
        "type": "object",
        "required": [
          "recipes"
        ],
        "title": "FeatureProcessingConfirmRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/feature_processing/confirm``.\n\nArgs:\n    recipes (dict[str, ColumnRecipe]): Final accepted recipes (accepted columns\n        only) \u2014 becomes this run's canonical ``feature_processing_recipes``.\n    dropped_columns (list[str]): Columns explicitly chosen \"Drop\" \u2014 removed\n        entirely, never reaching ``FeatureTransformChain``."
      },
      "FeatureProcessingConfirmResponse": {
        "properties": {
          "confirmed": {
            "type": "boolean",
            "title": "Confirmed"
          },
          "n_accepted": {
            "type": "integer",
            "title": "N Accepted"
          },
          "n_dropped": {
            "type": "integer",
            "title": "N Dropped"
          },
          "n_rejected": {
            "type": "integer",
            "title": "N Rejected"
          }
        },
        "type": "object",
        "required": [
          "confirmed",
          "n_accepted",
          "n_dropped",
          "n_rejected"
        ],
        "title": "FeatureProcessingConfirmResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/feature_processing/confirm``.\n\nArgs:\n    confirmed (bool): Always ``True`` on success.\n    n_accepted (int): Number of columns with a non-empty accepted recipe.\n    n_dropped (int): Number of columns dropped entirely.\n    n_rejected (int): Number of columns left as raw passthrough (present in the\n        candidate set but absent from both ``recipes`` and ``dropped_columns``)."
      },
      "FeatureProcessingPreviewResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/FeatureProcessingCandidate"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "candidates"
        ],
        "title": "FeatureProcessingPreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/feature_processing/preview``.\n\nArgs:\n    candidates (list[FeatureProcessingCandidate]): One entry per feature column\n        eligible for a recipe decision \u2014 excludes the date column (Step 3),\n        identifier columns (Step 2), and any boolean/datetime-dtype column (out\n        of scope for recipe treatment, matching ``DateFeatureExtractor`` being a\n        separate mechanism). Ordered \"Needing Attention\" columns first, then\n        \"Rest\", matching ``interactive_step_feature_processing.py``'s grouping."
      },
      "FeatureProcessingTransformPreviewRequest": {
        "properties": {
          "recipes": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ColumnRecipe"
            },
            "type": "object",
            "title": "Recipes"
          }
        },
        "type": "object",
        "required": [
          "recipes"
        ],
        "title": "FeatureProcessingTransformPreviewRequest",
        "description": "Request for ``POST /api/runs/{run_id}/steps/feature_processing/transform-preview``.\n\nArgs:\n    recipes (dict[str, ColumnRecipe]): The client's *current* (possibly edited)\n        accepted recipes \u2014 only columns the user has set to \"Accept\" belong here,\n        matching ``_render_live_preview()``'s ``accepted`` dict."
      },
      "FeatureProcessingTransformPreviewResponse": {
        "properties": {
          "before": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Before"
          },
          "after": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "After"
          },
          "columns_before": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns Before"
          },
          "columns_after": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Columns After"
          }
        },
        "type": "object",
        "required": [
          "before",
          "after",
          "columns_before",
          "columns_after"
        ],
        "title": "FeatureProcessingTransformPreviewResponse",
        "description": "Response for ``POST /api/runs/{run_id}/steps/feature_processing/transform-preview``.\n\nArgs:\n    before (list[dict[str, Any]]): First 5 training rows, accepted columns only,\n        JSON-safe records.\n    after (list[dict[str, Any]]): Same 5 rows after ``FeatureTransformChain``\n        fit_transform \u2014 column-expanding steps (e.g. one-hot) show their expanded\n        output columns, so ``after``'s keys may not match ``before``'s.\n    columns_before (list[str]): ``before``'s column order.\n    columns_after (list[str]): ``after``'s column order."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}